requirements.txt
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.5
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
blinker==1.6.2
build==0.10.0
cassandra-driver==3.27.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
executing==1.2.0
fanstatic==1.3
fastapi==0.95.2
fastjsonschema==2.17.1
Flask==2.3.2
fqdn==1.5.1
geomet==0.2.1.post1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
js==1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
lazy-object-proxy==1.9.0
make==0.1.6.post2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
mypy-extensions==0.4.3
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
orjson==3.8.14
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
pip-tools==6.13.0
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
Pygments==2.15.1
pylint==2.17.4
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pyscript==0.2.5
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==4.5.5
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rich-click==1.6.1
rsa==4.9
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
spotipy==2.23.0
SQLAlchemy==2.0.15
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
traitlets==5.9.0
typer==0.4.2
typing_extensions==4.6.2
ujson==5.7.0
uri-template==1.2.0
urllib3==2.0.2
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wrapt==1.15.0
